+++ lib/src/ui/components/logger/log_console_button.dart
@@ -10,8 +10,8 @@ class LogConsoleButton extends StatelessWidget implements ControlPanelSetting {
   Widget build(BuildContext context) {
     return Container(
       margin: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
-      child: RaisedButton(
-        color: Colors.green.withAlpha(240),
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(primary: Colors.green.withAlpha(240)),
         child: FittedBox(
           child: Text(
             'Show logs',