+++ lib/src/ui/components/logger/log_console_modal.dart
@@ -15,7 +15,7 @@ class LogConsoleModal extends ModalRoute<void> {
   Color get barrierColor => Colors.black.withOpacity(0.5);
 
   @override
-  String get barrierLabel => null;
+  String get barrierLabel => '';
 
   @override
   bool get maintainState => true;