+++ lib/src/ui/components/network/network_setting.dart
@@ -3,13 +3,13 @@ import 'package:flutter/material.dart';
 import 'package:dash_kit_control_panel/src/panel/control_panel_setting.dart';
 
 class NetworkSettingProps {
-  NetworkSettingProps({this.alice});
+  NetworkSettingProps({required this.alice});
 
   final Alice alice;
 }
 
 class NetworkSetting extends StatelessWidget implements ControlPanelSetting {
-  const NetworkSetting(this.props, {Key key}) : super(key: key);
+  const NetworkSetting(this.props, {Key? key}) : super(key: key);
 
   final NetworkSettingProps props;
 
@@ -20,8 +20,8 @@ class NetworkSetting extends StatelessWidget implements ControlPanelSetting {
   Widget build(BuildContext context) {
     return Container(
       margin: const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
-      child: RaisedButton(
-        color: Colors.green.withAlpha(240),
+      child: ElevatedButton(
+        style: ElevatedButton.styleFrom(primary: Colors.green.withAlpha(240)),
         child: FittedBox(
           child: Text(
             'Show network activity',