+++ lib/src/ui/control_panel_gate.dart
@@ -3,12 +3,12 @@ import 'package:dash_kit_control_panel/src/panel/control_panel.dart';
 
 class ControlPanelGate extends StatelessWidget {
   const ControlPanelGate({
-    Key key,
-    @required this.child,
+    Key? key,
+    required this.child,
     this.isEnabled = true,
   }) : super(key: key);
 
-  static DateTime _lastDoubleTapEvent;
+  static DateTime? _lastDoubleTapEvent;
 
   final Widget child;
   final bool isEnabled;
@@ -26,7 +26,7 @@ class ControlPanelGate extends StatelessWidget {
     final now = DateTime.now();
     final diff = _lastDoubleTapEvent != null
         ? now.millisecondsSinceEpoch -
-            _lastDoubleTapEvent.millisecondsSinceEpoch
+            _lastDoubleTapEvent!.millisecondsSinceEpoch
         : -1;
 
     if (_lastDoubleTapEvent != null && diff >= 0 && diff <= 2500) {