+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [3.0.0-dev.6]
+
+* Updated Readme file and example project
+
+## [3.0.0-dev.5]
+
+* Fixed an issue with device preview mode
+
+## [3.0.0-dev.4]
+
+* Fixed an issue with push token type
+
+## [3.0.0-dev.3]
+
+* Fixed issue with opening the panel and returned back device preview module
+
+## [3.0.0-dev.2]
+
+* Upgraded dependencies
+
+## [3.0.0-dev.1]
+
+* Added null-safety
+
 ## [1.0.1]
 
 * Upgraded dependencies